feat(wordle): wordsRemaining on game load
