+++ src/pages/wordle/context/reducer/helpers/index.js
@@ -1,3 +1,5 @@
+export * from "./getEligibleWords";
 export * from "./getLetterStatuses";
+export * from "./saveGame";
 export * from "./updateBoard";
 export * from "./updateLetterStatuses";
